+++ src/scratchpad/src/org/apache/poi/hemf/record/emfplus/HemfPlusObject.java
@@ -265,7 +265,7 @@ public class HemfPlusObject {
 
             long size = graphicsVersion.init(leis);
 
-            objectDataBytes = IOUtils.toByteArray(leis, dataSize - size, MAX_OBJECT_SIZE);
+            objectDataBytes = IOUtils.toByteArray(leis, (int)(dataSize - size), MAX_OBJECT_SIZE);
 
             return dataSize;
         }
